+++ example/android/app/src/main/kotlin/fr/skyost/bubble_showcase_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package fr.skyost.bubble_showcase_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}