+++ lib/src/showcase.dart
@@ -23,18 +23,18 @@ class BubbleShowcase extends StatefulWidget {
   final Widget child;
 
   /// The counter text (:i is the current slide, :n is the slides count). You can pass null to disable this.
-  final String counterText;
+  final String? counterText;
 
   /// Whether to show a close button.
   final bool showCloseButton;
 
   /// Creates a new bubble showcase instance.
   BubbleShowcase({
-    @required this.bubbleShowcaseId,
-    @required this.bubbleShowcaseVersion,
+    required this.bubbleShowcaseId,
+    required this.bubbleShowcaseVersion,
     this.doNotReopenOnClose = false,
-    @required this.bubbleSlides,
-    this.child,
+    required this.bubbleSlides,
+    required this.child,
     this.counterText = ':i/:n',
     this.showCloseButton = true,
   }) : assert(bubbleSlides.isNotEmpty);
@@ -48,31 +48,27 @@ class BubbleShowcase extends StatefulWidget {
       return true;
     }
     SharedPreferences preferences = await SharedPreferences.getInstance();
-    bool result =
-        preferences.getBool('$bubbleShowcaseId.$bubbleShowcaseVersion');
+    bool? result = preferences.getBool('$bubbleShowcaseId.$bubbleShowcaseVersion');
     return result == null || result;
   }
 }
 
 /// The BubbleShowcase state.
-class _BubbleShowcaseState extends State<BubbleShowcase>
-    with WidgetsBindingObserver {
+class _BubbleShowcaseState extends State<BubbleShowcase> with WidgetsBindingObserver {
   /// The current slide index.
-  int _currentSlideIndex = -1;
+  int currentSlideIndex = -1;
 
   /// The current slide entry.
-  OverlayEntry _currentSlideEntry;
+  OverlayEntry? currentSlideEntry;
 
   @override
   void initState() {
-    WidgetsBinding.instance.addPostFrameCallback((_) async {
+    WidgetsBinding.instance?.addPostFrameCallback((_) async {
       if (await widget.shouldOpenShowcase) {
-        _currentSlideIndex++;
-        _currentSlideEntry = _createCurrentSlideEntry();
-        Overlay.of(context).insert(_currentSlideEntry);
+        goToNextEntryOrClose(0);
       }
     });
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance?.addObserver(this);
 
     super.initState();
   }
@@ -82,57 +78,73 @@ class _BubbleShowcaseState extends State<BubbleShowcase>
 
   @override
   void dispose() {
-    _currentSlideEntry?.remove();
-    WidgetsBinding.instance.removeObserver(this);
+    currentSlideEntry?.remove();
+    WidgetsBinding.instance?.removeObserver(this);
     super.dispose();
   }
 
   @override
   void didChangeMetrics() {
-    if (_currentSlideEntry == null) {
-      return;
-    }
-
-    SchedulerBinding.instance.addPostFrameCallback((_) {
-      _currentSlideEntry.remove();
-      Overlay.of(context).insert(_currentSlideEntry);
+    SchedulerBinding.instance?.addPostFrameCallback((_) {
+      if (currentSlideEntry != null) {
+        currentSlideEntry!.remove();
+        Overlay.of(context)?.insert(currentSlideEntry!);
+      }
     });
   }
 
   /// Returns whether the showcasing is finished.
-  bool get _isFinished =>
-      _currentSlideIndex == -1 ||
-      _currentSlideIndex == widget.bubbleSlides.length;
+  bool get isFinished => currentSlideIndex == -1 || currentSlideIndex == widget.bubbleSlides.length;
 
   /// Allows to go to the next entry (or to close the showcase if needed).
-  void _goToNextEntryOrClose(int position) {
-    _currentSlideIndex = position;
-    _currentSlideEntry.remove();
+  void goToNextEntryOrClose(int position) {
+    currentSlideIndex = position;
+    currentSlideEntry?.remove();
+    triggerOnExit();
 
-    if (_isFinished) {
-      _currentSlideEntry = null;
+    if (isFinished) {
+      currentSlideEntry = null;
       if (widget.doNotReopenOnClose) {
         SharedPreferences.getInstance().then((preferences) {
-          preferences.setBool(
-              '${widget.bubbleShowcaseId}.${widget.bubbleShowcaseVersion}',
-              false);
+          preferences.setBool('${widget.bubbleShowcaseId}.${widget.bubbleShowcaseVersion}', false);
         });
       }
     } else {
-      _currentSlideEntry = _createCurrentSlideEntry();
-      Overlay.of(context).insert(_currentSlideEntry);
+      currentSlideEntry = createCurrentSlideEntry();
+      Overlay.of(context)?.insert(currentSlideEntry!);
+      triggerOnEnter();
     }
   }
 
   /// Creates the current slide entry.
-  OverlayEntry _createCurrentSlideEntry() => OverlayEntry(
-        builder: (context) => widget.bubbleSlides[_currentSlideIndex].build(
+  OverlayEntry createCurrentSlideEntry() => OverlayEntry(
+        builder: (context) => widget.bubbleSlides[currentSlideIndex].build(
           context,
           widget,
-          _currentSlideIndex,
+          currentSlideIndex,
           (position) {
-            setState(() => _goToNextEntryOrClose(position));
+            setState(() => goToNextEntryOrClose(position));
           },
         ),
       );
+
+  /// Allows to trigger enter callbacks.
+  void triggerOnEnter() {
+    if (currentSlideIndex >= 0 && currentSlideIndex < widget.bubbleSlides.length) {
+      VoidCallback? callback = widget.bubbleSlides[currentSlideIndex].onEnter;
+      if (callback != null) {
+        callback();
+      }
+    }
+  }
+
+  /// Allows to trigger exit callbacks.
+  void triggerOnExit() {
+    if (currentSlideIndex >= 0 && currentSlideIndex < widget.bubbleSlides.length) {
+      VoidCallback? callback = widget.bubbleSlides[currentSlideIndex].onExit;
+      if (callback != null) {
+        callback();
+      }
+    }
+  }
 }