+++ lib/src/slide.dart
@@ -15,8 +15,14 @@ abstract class BubbleSlide {
   /// The box shadow.
   final BoxShadow boxShadow;
 
+  /// Triggered when this slide has been entered.
+  final VoidCallback? onEnter;
+
+  /// Triggered when this slide has been exited.
+  final VoidCallback? onExit;
+
   /// The slide child.
-  final BubbleSlideChild child;
+  final BubbleSlideChild? child;
 
   /// Creates a new bubble slide instance.
   const BubbleSlide({
@@ -26,14 +32,14 @@ abstract class BubbleSlide {
       blurRadius: 0,
       spreadRadius: 0,
     ),
+    this.onEnter,
+    this.onExit,
     this.child,
   });
 
   /// Builds the whole slide widget.
-  Widget build(BuildContext context, BubbleShowcase bubbleShowcase,
-      int currentSlideIndex, void Function(int) goToSlide) {
-    Position highlightPosition =
-        getHighlightPosition(context, bubbleShowcase, currentSlideIndex);
+  Widget build(BuildContext context, BubbleShowcase bubbleShowcase, int currentSlideIndex, void Function(int) goToSlide) {
+    Position highlightPosition = getHighlightPosition(context, bubbleShowcase, currentSlideIndex);
     List<Widget> children = [
       Positioned.fill(
         child: CustomPaint(
@@ -42,14 +48,12 @@ abstract class BubbleSlide {
       ),
     ];
 
-    if (child != null && child.widget != null) {
-      children.add(
-          child.build(context, highlightPosition, MediaQuery.of(context).size));
+    if (child?.widget != null) {
+      children.add(child!.build(context, highlightPosition, MediaQuery.of(context).size));
     }
 
     int slidesCount = bubbleShowcase.bubbleSlides.length;
-    Color writeColor =
-        Utils.isColorDark(boxShadow.color) ? Colors.white : Colors.black;
+    Color writeColor = Utils.isColorDark(boxShadow.color) ? Colors.white : Colors.black;
     if (bubbleShowcase.counterText != null) {
       children.add(
         Positioned(
@@ -57,11 +61,8 @@ abstract class BubbleSlide {
           left: 0,
           right: 0,
           child: Text(
-            bubbleShowcase.counterText
-                .replaceAll(':i', (currentSlideIndex + 1).toString())
-                .replaceAll(':n', slidesCount.toString()),
-            style:
-                Theme.of(context).textTheme.body1.copyWith(color: writeColor),
+            bubbleShowcase.counterText!.replaceAll(':i', (currentSlideIndex + 1).toString()).replaceAll(':n', slidesCount.toString()),
+            style: Theme.of(context).textTheme.bodyText2!.copyWith(color: writeColor),
             textAlign: TextAlign.center,
           ),
         ),
@@ -73,11 +74,11 @@ abstract class BubbleSlide {
         top: MediaQuery.of(context).padding.top,
         left: 0,
         child: GestureDetector(
+          onTap: () => goToSlide(slidesCount),
           child: Icon(
             Icons.close,
             color: writeColor,
           ),
-          onTap: () => goToSlide(slidesCount),
         ),
       ));
     }
@@ -91,8 +92,7 @@ abstract class BubbleSlide {
   }
 
   /// Returns the position to highlight.
-  Position getHighlightPosition(BuildContext context,
-      BubbleShowcase bubbleShowcase, int currentSlideIndex);
+  Position getHighlightPosition(BuildContext context, BubbleShowcase bubbleShowcase, int currentSlideIndex);
 }
 
 /// A bubble slide with a position that depends on another widget.
@@ -108,8 +108,8 @@ class RelativeBubbleSlide extends BubbleSlide {
       blurRadius: 0,
       spreadRadius: 0,
     ),
-    BubbleSlideChild child,
-    @required this.widgetKey,
+    required BubbleSlideChild child,
+    required this.widgetKey,
   }) : super(
           shape: shape,
           boxShadow: boxShadow,
@@ -117,10 +117,8 @@ class RelativeBubbleSlide extends BubbleSlide {
         );
 
   @override
-  Position getHighlightPosition(BuildContext context,
-      BubbleShowcase bubbleShowcase, int currentSlideIndex) {
-    RenderBox renderBox =
-        widgetKey.currentContext.findRenderObject() as RenderBox;
+  Position getHighlightPosition(BuildContext context, BubbleShowcase bubbleShowcase, int currentSlideIndex) {
+    RenderBox renderBox = widgetKey.currentContext!.findRenderObject() as RenderBox;
     Offset offset = renderBox.localToGlobal(Offset.zero);
 
     return Position(
@@ -145,8 +143,8 @@ class AbsoluteBubbleSlide extends BubbleSlide {
       blurRadius: 0,
       spreadRadius: 0,
     ),
-    BubbleSlideChild child,
-    @required this.positionCalculator,
+    required BubbleSlideChild child,
+    required this.positionCalculator,
   }) : super(
           shape: shape,
           boxShadow: boxShadow,
@@ -154,9 +152,7 @@ class AbsoluteBubbleSlide extends BubbleSlide {
         );
 
   @override
-  Position getHighlightPosition(BuildContext context,
-          BubbleShowcase bubbleShowcase, int currentSlideIndex) =>
-      positionCalculator(MediaQuery.of(context).size);
+  Position getHighlightPosition(BuildContext context, BubbleShowcase bubbleShowcase, int currentSlideIndex) => positionCalculator(MediaQuery.of(context).size);
 }
 
 /// A bubble slide child, holding a widget.
@@ -166,7 +162,7 @@ abstract class BubbleSlideChild {
 
   /// Creates a new bubble slide child instance.
   const BubbleSlideChild({
-    this.widget,
+    required this.widget,
   });
 
   /// Builds the bubble slide child widget.
@@ -182,8 +178,7 @@ abstract class BubbleSlideChild {
   }
 
   /// Returns child position according to the highlight position and parent size.
-  Position getPosition(
-      BuildContext context, Position highlightPosition, Size parentSize);
+  Position getPosition(BuildContext context, Position highlightPosition, Size parentSize);
 }
 
 /// A bubble slide with a position that depends on the highlight zone.
@@ -193,15 +188,14 @@ class RelativeBubbleSlideChild extends BubbleSlideChild {
 
   /// Creates a new relative bubble slide child instance.
   const RelativeBubbleSlideChild({
-    Widget widget,
+    required Widget widget,
     this.direction = AxisDirection.down,
   }) : super(
           widget: widget,
         );
 
   @override
-  Position getPosition(
-      BuildContext context, Position highlightPosition, Size parentSize) {
+  Position getPosition(BuildContext context, Position highlightPosition, Size parentSize) {
     switch (direction) {
       case AxisDirection.up:
         return Position(
@@ -238,14 +232,12 @@ class AbsoluteBubbleSlideChild extends BubbleSlideChild {
 
   /// Creates a new absolute bubble slide child instance.
   const AbsoluteBubbleSlideChild({
-    Widget widget,
-    @required this.positionCalculator,
+    required Widget widget,
+    required this.positionCalculator,
   }) : super(
           widget: widget,
         );
 
   @override
-  Position getPosition(
-          BuildContext context, Position highlightPosition, Size parentSize) =>
-      positionCalculator(parentSize);
+  Position getPosition(BuildContext context, Position highlightPosition, Size parentSize) => positionCalculator(parentSize);
 }